+++ packages-next/admin-ui/src/templates/app.ts
@@ -20,14 +20,15 @@ type AppTemplateOptions = {
 };
 
 export const appTemplate = (
-  keystone: KeystoneSystem,
+  system: KeystoneSystem,
   { configFile, projectAdminPath }: AppTemplateOptions
 ) => {
-  const lazyMetadataQuery = getLazyMetadataQuery(keystone.graphQLSchema, keystone.adminMeta);
+  const { graphQLSchema, adminMeta, views } = system;
+  const lazyMetadataQuery = getLazyMetadataQuery(graphQLSchema, adminMeta);
 
   const result = executeSync({
     document: staticAdminMetaQuery,
-    schema: keystone.graphQLSchema,
+    schema: graphQLSchema,
     contextValue: {
       isAdminUIBuildProcess: true,
     },
@@ -44,7 +45,7 @@ import { KeystoneProvider } from '@keystone-next/admin-ui/context';
 import { ErrorBoundary } from '@keystone-next/admin-ui/components';
 import { Core } from '@keystone-ui/core';
 
-${keystone.views
+${views
   .map(
     (view, i) =>
       `import * as view${i} from ${JSON.stringify(
@@ -55,7 +56,7 @@ ${keystone.views
 
 ${configFile ? `import * as adminConfig from "../../../admin/config";` : 'const adminConfig = {};'}
 
-const fieldViews = [${keystone.views.map((x, i) => `view${i}`)}];
+const fieldViews = [${views.map((x, i) => `view${i}`)}];
 
 const lazyMetadataQuery = ${JSON.stringify(lazyMetadataQuery)};