+++ packages-next/admin-ui/src/templates/index.ts
@@ -13,10 +13,12 @@ const pkgDir = Path.dirname(require.resolve('@keystone-next/admin-ui/package.jso
 export { adminMetaSchemaExtension } from './adminMetaSchemaExtension';
 
 export const writeAdminFiles = (
-  keystone: KeystoneSystem,
+  system: KeystoneSystem,
   configFile: boolean,
   projectAdminPath: string
 ): AdminFileToWrite[] => {
+  const { adminMeta } = system;
+  const { session } = system.config;
   return [
     {
       mode: 'copy',
@@ -31,30 +33,30 @@ export const writeAdminFiles = (
     {
       mode: 'write',
       outputPath: 'pages/no-access.js',
-      src: noAccessTemplate(keystone),
+      src: noAccessTemplate(session),
     },
     {
       mode: 'write',
       outputPath: 'pages/_app.js',
-      src: appTemplate(keystone, { configFile, projectAdminPath }),
+      src: appTemplate(system, { configFile, projectAdminPath }),
     },
     {
       mode: 'write',
-      src: homeTemplate(keystone),
+      src: homeTemplate(adminMeta),
       outputPath: 'pages/index.js',
     },
-    ...Object.values(keystone.adminMeta.lists)
+    ...Object.values(adminMeta.lists)
       .map(list => {
         const { path } = list;
         return [
           {
             mode: 'write',
-            src: listTemplate(keystone, { list }),
+            src: listTemplate({ list }),
             outputPath: `pages/${path}/index.js`,
           },
           {
             mode: 'write',
-            src: itemTemplate(keystone, { list }),
+            src: itemTemplate({ list }),
             outputPath: `pages/${path}/[id].js`,
           },
         ] as const;