+++ packages-next/admin-ui/src/templates/no-access.ts
@@ -1,6 +1,6 @@
-import type { KeystoneSystem } from '@keystone-next/types';
+import type { KeystoneConfig } from '@keystone-next/types';
 
-export const noAccessTemplate = (keystone: KeystoneSystem) => {
+export const noAccessTemplate = (session: KeystoneConfig['session']) => {
   // -- TEMPLATE START
   return `
 import React from 'react';
@@ -8,7 +8,7 @@ import React from 'react';
 import { NoAccessPage } from '@keystone-next/admin-ui/pages/NoAccessPage';
 
 export default function Home() {
-  return <NoAccessPage sessionsEnabled={${!!keystone.config.session}} />;
+  return <NoAccessPage sessionsEnabled={${!!session}} />;
 }
   `;
   // -- TEMPLATE END