+++ packages-next/auth/src/gql/getInitFirstItemSchema.ts
@@ -1,5 +1,3 @@
-import { GraphQLSchema } from 'graphql';
-import { graphQLSchemaExtension } from '@keystone-next/keystone/schema';
 import { AuthGqlNames } from '../types';
 
 export function getInitFirstItemSchema({
@@ -7,22 +5,21 @@ export function getInitFirstItemSchema({
   fields,
   itemData,
   gqlNames,
+  keystone,
 }: {
   listKey: string;
   fields: string[];
   itemData: Record<string, any> | undefined;
   gqlNames: AuthGqlNames;
+  keystone: any;
 }) {
-  return (schema: GraphQLSchema, keystoneClassInstance: any) => {
-    const list = keystoneClassInstance.lists[listKey];
-
-    const newSchema = graphQLSchemaExtension({
-      typeDefs: `
+  return {
+    typeDefs: `
         input ${gqlNames.CreateInitialInput} {
           ${Array.prototype
             .concat(
               ...fields.map(fieldPath =>
-                list.fieldsByPath[fieldPath].gqlCreateInputFields({
+                keystone.lists[listKey].fieldsByPath[fieldPath].gqlCreateInputFields({
                   schemaName: 'public',
                 })
               )
@@ -31,38 +28,22 @@ export function getInitFirstItemSchema({
         }
         type Mutation {
           ${gqlNames.createInitialItem}(data: ${gqlNames.CreateInitialInput}!): ${
-        gqlNames.ItemAuthenticationWithPasswordResult
-      }!
+      gqlNames.ItemAuthenticationWithPasswordSuccess
+    }!
         }
       `,
-      resolvers: {
-        Mutation: {
-          async [`createInitial${listKey}`](rootVal: any, { data }: any, context: any) {
-            const { count } = await list.adapter.itemsQuery(
-              {},
-              {
-                meta: true,
-              }
-            );
-            if (count !== 0) {
-              throw new Error('Initial items can only be created when no items exist in that list');
-            }
-            const contextThatSkipsAccessControl = await context.createContext({
-              skipAccessControl: true,
-            });
-            const item = await list.createMutation(
-              { ...data, ...itemData },
-              contextThatSkipsAccessControl
-            );
-            const token = await context.startSession({ listKey, itemId: item.id });
-            return {
-              item,
-              token,
-            };
-          },
+    resolvers: {
+      Mutation: {
+        async [gqlNames.createInitialItem](rootVal: any, { data }: any, context: any) {
+          const count = await context.lists[listKey].count({});
+          if (count !== 0) {
+            throw new Error('Initial items can only be created when no items exist in that list');
+          }
+          const item = await context.lists[listKey].createOne({ data: { ...data, ...itemData } });
+          const sessionToken = await context.startSession({ listKey, itemId: item.id });
+          return { item, sessionToken };
         },
       },
-    })(schema, keystoneClassInstance);
-    return newSchema;
+    },
   };
 }