+++ packages-next/keystone/src/lib/createAdminUIServer.ts
@@ -6,32 +6,61 @@ import express from 'express';
 import { ApolloServer } from 'apollo-server-express';
 // @ts-ignore
 import { formatError } from '@keystonejs/keystone/lib/Keystone/format-error';
-import type { KeystoneSystem } from '@keystone-next/types';
+import type { KeystoneSystem, KeystoneConfig } from '@keystone-next/types';
 
 const dev = process.env.NODE_ENV !== 'production';
 
-export const createAdminUIServer = async (keystone: KeystoneSystem) => {
+const addApolloServer = ({ server, system }: { server: any; system: KeystoneSystem }) => {
+  const { graphQLSchema, createContextFromRequest } = system;
+  const apolloServer = new ApolloServer({
+    // FIXME: Support for file handling configuration
+    // maxFileSize: 200 * 1024 * 1024,
+    // maxFiles: 5,
+    schema: graphQLSchema,
+    // FIXME: allow the dev to control where/when they get a playground
+    playground: { settings: { 'request.credentials': 'same-origin' } },
+    formatError, // TODO: this needs to be discussed
+    context: ({ req, res }) => createContextFromRequest(req, res),
+    // FIXME: support for apollo studio tracing
+    // ...(process.env.ENGINE_API_KEY || process.env.APOLLO_KEY
+    //   ? { tracing: true }
+    //   : {
+    //       engine: false,
+    //       // Only enable tracing in dev mode so we can get local debug info, but
+    //       // don't bother returning that info on prod when the `engine` is
+    //       // disabled.
+    //       tracing: dev,
+    //     }),
+    // FIXME: Support for generic custom apollo configuration
+    // ...apolloConfig,
+  });
+  // FIXME: Support custom API path via config.graphql.path.
+  // Note: Core keystone uses '/admin/api' as the default.
+  apolloServer.applyMiddleware({ app: server, path: '/api/graphql', cors: false });
+};
+
+export const createAdminUIServer = async (config: KeystoneConfig, system: KeystoneSystem) => {
   const server = express();
 
-  // TODO: allow cors to be configured
-  server.use(cors({ origin: true, credentials: true }));
+  if (config.server?.cors) {
+    // Setting config.server.cors = true will provide backwards compatible defaults
+    // Otherwise, the user can provide their own config object to use
+    const corsConfig =
+      typeof config.server.cors === 'boolean'
+        ? { origin: true, credentials: true }
+        : config.server.cors;
+    server.use(cors(corsConfig));
+  }
 
   console.log('✨ Preparing Next.js app');
   const app = next({ dev, dir: Path.join(process.cwd(), '.keystone', 'admin') });
   const handle = app.getRequestHandler();
-  await Promise.all([app.prepare(), keystone.keystone.connect()]);
+  await Promise.all([app.prepare(), system.keystone.connect()]);
 
   console.log('✨ Preparing GraphQL Server');
-  const apolloServer = new ApolloServer({
-    schema: keystone.graphQLSchema,
-    playground: { settings: { 'request.credentials': 'same-origin' } },
-    // TODO: this needs to be discussed
-    formatError,
-    context: ({ req, res }) => keystone.createContextFromRequest(req, res),
-  });
-  apolloServer.applyMiddleware({ app: server, path: '/api/graphql' });
+  addApolloServer({ server, system });
 
-  const publicPages = keystone.config.ui?.publicPages ?? [];
+  const publicPages = system.config.ui?.publicPages ?? [];
 
   server.use(async (req, res) => {
     const { pathname } = url.parse(req.url);
@@ -39,15 +68,15 @@ export const createAdminUIServer = async (keystone: KeystoneSystem) => {
       handle(req, res);
       return;
     }
-    const session = (await keystone.createSessionContext?.(req, res))?.session;
-    const isValidSession = keystone.config.ui?.isAccessAllowed
-      ? await keystone.config.ui.isAccessAllowed({ session })
+    const session = (await system.createSessionContext?.(req, res))?.session;
+    const isValidSession = system.config.ui?.isAccessAllowed
+      ? await system.config.ui.isAccessAllowed({ session })
       : session !== undefined;
-    const maybeRedirect = await keystone.config.ui?.pageMiddleware?.({
+    const maybeRedirect = await system.config.ui?.pageMiddleware?.({
       req,
       session,
       isValidSession,
-      keystone,
+      system,
     });
     if (maybeRedirect) {
       res.redirect(maybeRedirect.to);