+++ packages-next/keystone/src/scripts/dev.ts
@@ -2,7 +2,7 @@ import path from 'path';
 import express from 'express';
 import { printSchema } from 'graphql';
 import * as fs from 'fs-extra';
-import { createKeystone } from '../lib/createKeystone';
+import { createSystem } from '../lib/createSystem';
 import { requireSource } from '../lib/requireSource';
 import { formatSource, generateAdminUI } from '../lib/generateAdminUI';
 import { createAdminUIServer } from '../lib/createAdminUIServer';
@@ -27,19 +27,20 @@ export const dev = async () => {
   let adminUIServer: null | ReturnType<typeof express> = null;
 
   const initKeystone = async () => {
-    const keystone = createKeystone(requireSource(path.join(process.cwd(), 'keystone')).default);
-    let printedSchema = printSchema(keystone.graphQLSchema);
+    const config = requireSource(path.join(process.cwd(), 'keystone')).default;
+    const system = createSystem(config);
+    let printedSchema = printSchema(system.graphQLSchema);
     console.log('✨ Generating Schema');
     await fs.outputFile('./.keystone/schema.graphql', printedSchema);
     await fs.outputFile(
       './.keystone/schema-types.ts',
-      formatSource(printGeneratedTypes(printedSchema, keystone), 'babel-ts')
+      formatSource(printGeneratedTypes(printedSchema, system), 'babel-ts')
     );
 
     console.log('✨ Generating Admin UI');
-    await generateAdminUI(keystone, process.cwd());
+    await generateAdminUI(system, process.cwd());
 
-    adminUIServer = await createAdminUIServer(keystone);
+    adminUIServer = await createAdminUIServer(config, system);
     console.log(`👋 Admin UI Ready`);
   };