+++ packages-next/fields/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @keystone-next/fields
 
+## 3.0.1
+
+### Patch Changes
+
+- [`81a140ee3`](https://github.com/keystonejs/keystone/commit/81a140ee3badc9c032ab02a233a21d011278e173) [#4289](https://github.com/keystonejs/keystone/pull/4289) Thanks [@JedWatson](https://github.com/JedWatson)! - Fixed many relationship select clipping in the Create drawer
+
+- Updated dependencies [[`11777cddb`](https://github.com/keystonejs/keystone/commit/11777cddba45b28a9e17a3149b792db121322b46), [`cbf11a69b`](https://github.com/keystonejs/keystone/commit/cbf11a69b8f2c428e2c0a08dd568b3bc0e0d80f4), [`b2de22941`](https://github.com/keystonejs/keystone/commit/b2de229419cc93b69ee4027c387cab9c8d701488), [`cc987d078`](https://github.com/keystonejs/keystone/commit/cc987d078653fd9e686069f9f885f1269b64a882), [`60e061246`](https://github.com/keystonejs/keystone/commit/60e061246bc35b76031f43ff6c07446fe6ad3c6b), [`cc987d078`](https://github.com/keystonejs/keystone/commit/cc987d078653fd9e686069f9f885f1269b64a882)]:
+  - @keystone-next/admin-ui@3.0.0
+  - @keystone-next/types@4.0.0
+  - @keystone-next/admin-ui-utils@1.0.2
+
 ## 3.0.0
 
 ### Major Changes